backend/pyproject.toml
[project]
name = "My_Tasks_Backend"
version = "0.1.0"
description = "My Task Management System"
authors = [{ name = "Michał Łowkiet", email = "michallowkiet@gmail.com" }]
license = "MIT"

readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "djoser>=2.3.1",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
select = ["F", "W", "E", "C", "DJ"]

exclude = [
    ".git",
    "__pycache__",
    "*.pyc",
    "*.egg-info",
    "*.dist-info",
    "venv",
    "build",
    "dist",
]

line-length = 120
